requirements.txt
google-generativeai>=0.1.0
requests>=2.28.0
python-docx>=0.8.11
pdfplumber>=0.9.0
python-dotenv>=1.0.0
